# specs/001-project-aether/contracts/api.yaml
openapi: 3.1.0
info:
  title: Project Aether API
  description: 'API for Project Aether — an intelligent home automation system with multi-agent architecture.


    ## Architecture


    Aether uses a multi-agent system orchestrated via LangGraph workflows and traced via MLflow. An Orchestrator agent classifies
    user intent, selects the appropriate domain agent (Architect, Data Science team, Food, Research, Knowledge, Dashboard Designer),
    and dynamically configures each agent with the right model, tools, and prompt. Mutating actions require HITL approval via
    the web UI or push notifications (iPhone/Apple Watch).


    ## Constitution Compliance


    - **Safety First**: All automation proposals require explicit HITL approval before deployment

    - **Isolation**: All analysis scripts run in sandboxed containers (gVisor/Podman)

    - **Observability**: Every agent interaction is traced via MLflow

    - **State**: LangGraph + PostgreSQL for persistent state management


    ## OpenAI Compatibility


    The /v1/chat/completions and /v1/models endpoints are OpenAI-compatible, allowing integration with tools like Open WebUI.

    '
  version: 0.4.0
  contact:
    name: Project Aether
  license:
    name: MIT
paths:
...
  /api/v1/entities:
    get:
      tags:
      - Entities
      summary: List Entities
      description: List all entities with optional filtering.
      operationId: list_entities_api_v1_entities_get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: domain
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by domain
          title: Domain
        description: Filter by domain
      - name: area_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by area
          title: Area Id
        description: Filter by area
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by state
          title: State
        description: Filter by state
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/entities/{entity_id}:
    get:
      tags:
      - Entities
      summary: Get Entity
      description: Get a specific entity by HA entity_id.
      operationId: get_entity_api_v1_entities__entity_id__get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/entities/query:
    post:
      tags:
      - Entities
      summary: Query Entities
      description: 'Query entities using natural language.


        Rate limited to 10/minute (LLM-backed).


        Examples:

        - "all lights in the living room"

        - "temperature sensors"

        - "devices that are on"'
      operationId: query_entities_api_v1_entities_query_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityQueryResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
  /api/v1/entities/sync:
    post:
      tags:
      - Entities
      summary: Sync Entities
      description: 'Trigger entity discovery and sync from Home Assistant.


        Rate limited to 5/minute (expensive MCP + DB operation).'
      operationId: sync_entities_api_v1_entities_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntitySyncRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitySyncResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
  /api/v1/entities/domains/summary:
    get:
      tags:
      - Entities
      summary: Get Domain Summary
      description: Get entity count per domain.
      operationId: get_domain_summary_api_v1_entities_domains_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: integer
                type: object
                title: Response Get Domain Summary Api V1 Entities Domains Summary Get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
...
  /api/v1/usage/summary:
    get:
      tags:
      - Usage
      summary: Get Usage Summary
      description: 'Get LLM usage summary for the specified period.


        Returns total calls, tokens, cost, and per-model breakdown.'
      operationId: get_usage_summary_api_v1_usage_summary_get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Number of days to summarize
          default: 30
          title: Days
        description: Number of days to summarize
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Usage Summary Api V1 Usage Summary Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/usage/daily:
    get:
      tags:
      - Usage
      summary: Get Daily Usage
      description: 'Get daily LLM usage breakdown for charting.


        Returns a list of daily entries with calls, tokens, and cost.'
      operationId: get_daily_usage_api_v1_usage_daily_get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Number of days of history
          default: 30
          title: Days
        description: Number of days of history
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Daily Usage Api V1 Usage Daily Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/usage/models:
    get:
      tags:
      - Usage
      summary: Get Usage By Model
      description: 'Get per-model LLM usage breakdown.


        Returns calls, tokens, cost, and average latency for each model.'
      operationId: get_usage_by_model_api_v1_usage_models_get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          description: Number of days to analyze
          default: 30
          title: Days
        description: Number of days to analyze
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Usage By Model Api V1 Usage Models Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/usage/conversation/{conversation_id}:
    get:
      tags:
      - Usage
      summary: Get Conversation Cost
      description: 'Get aggregated cost and usage for a specific conversation.


        Returns total calls, tokens, cost, and per-agent breakdown.'
      operationId: get_conversation_cost_api_v1_usage_conversation__conversation_id__get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Conversation Cost Api V1 Usage Conversation  Conversation Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/models/summary:
    get:
      tags:
      - Model Registry
      summary: Model Summary
      description: Get aggregated model summaries with average ratings.
      operationId: model_summary_api_v1_models_summary_get
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
      parameters:
      - name: agent_role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Role
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelSummaryResponse'
                title: Response Model Summary Api V1 Models Summary Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/models:
    get:
      tags:
      - OpenAI Compatible
      summary: List Models
      description: 'List available LLM models.


        Dynamically discovers available models from:

        - Ollama (local models - if running)

        - Configured provider (openrouter, openai, google)


        Results are cached for 5 minutes.

        All models power the Architect agent with Home Assistant tools.'
      operationId: list_models_api_v1_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsResponse'
      security:
      - APIKeyHeader: []
      - APIKeyQuery: []
...
components:
  schemas:
...
    EntityListResponse:
      properties:
        entities:
          items:
            $ref: '#/components/schemas/EntityResponse'
          type: array
          title: Entities
        total:
          type: integer
          title: Total
        domain:
          anyOf:
          - type: string
          - type: 'null'
          title: Domain
      type: object
      required:
      - entities
      - total
      title: EntityListResponse
      description: Response for entity list.
    EntityQueryRequest:
      properties:
        query:
          type: string
          maxLength: 2000
          title: Query
          description: Natural language query
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          default: 20
      type: object
      required:
      - query
      title: EntityQueryRequest
      description: Request for natural language entity query.
    EntityQueryResult:
      properties:
        entities:
          items:
            $ref: '#/components/schemas/EntityResponse'
          type: array
          title: Entities
        query:
          type: string
          title: Query
        interpreted_as:
          anyOf:
          - type: string
          - type: 'null'
          title: Interpreted As
      type: object
      required:
      - entities
      - query
      title: EntityQueryResult
      description: Result from entity query.
    EntityResponse:
      properties:
        entity_id:
          type: string
          title: Entity Id
          description: HA entity ID (e.g., light.living_room)
        domain:
          type: string
          title: Domain
          description: Entity domain
        name:
          type: string
          title: Name
          description: Display name
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
          description: Current state
        id:
          type: string
          title: Id
          description: Internal UUID
        attributes:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Attributes
          description: Entity attributes
        area_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Id
          description: Area ID
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
          description: Device ID
        device_class:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Class
        unit_of_measurement:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Of Measurement
        supported_features:
          type: integer
          title: Supported Features
          default: 0
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        last_synced_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Synced At
      type: object
      required:
      - entity_id
      - domain
      - name
      - id
      title: EntityResponse
      description: Entity response with all fields.
    EntitySyncRequest:
      properties:
        force:
          type: boolean
          title: Force
          description: Force full re-sync
          default: false
      type: object
      title: EntitySyncRequest
      description: Request to trigger entity sync.
    EntitySyncResponse:
      properties:
        session_id:
          type: string
          title: Session Id
        status:
          type: string
          title: Status
        entities_found:
          type: integer
          title: Entities Found
        entities_added:
          type: integer
          title: Entities Added
        entities_updated:
          type: integer
          title: Entities Updated
        entities_removed:
          type: integer
          title: Entities Removed
        duration_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration Seconds
      type: object
      required:
      - session_id
      - status
      - entities_found
      - entities_added
      - entities_updated
      - entities_removed
      title: EntitySyncResponse
      description: Response from entity sync.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModelInfo:
      properties:
        id:
          type: string
          title: Id
        object:
          type: string
          title: Object
          default: model
        created:
          type: integer
          title: Created
        owned_by:
          type: string
          title: Owned By
          default: aether
        input_cost_per_1m:
          anyOf:
          - type: number
          - type: 'null'
          title: Input Cost Per 1M
        output_cost_per_1m:
          anyOf:
          - type: number
          - type: 'null'
          title: Output Cost Per 1M
      type: object
      required:
      - id
      title: ModelInfo
      description: Model information.
...
    ModelSummaryResponse:
      properties:
        model_name:
          type: string
          title: Model Name
        agent_role:
          type: string
          title: Agent Role
        avg_rating:
          type: number
          title: Avg Rating
        rating_count:
          type: integer
          title: Rating Count
        latest_config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Latest Config
      type: object
      required:
      - model_name
      - agent_role
      - avg_rating
      - rating_count
      - latest_config
      title: ModelSummaryResponse
      description: Aggregated model summary with average rating.
    ModelsResponse:
      properties:
        object:
          type: string
          title: Object
          default: list
        data:
          items:
            $ref: '#/components/schemas/ModelInfo'
          type: array
          title: Data
      type: object
      required:
      - data
      title: ModelsResponse
      description: List models response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    APIKeyQuery:
      type: apiKey
      in: query
      name: api_key
servers:
- url: http://localhost:8000
  description: Local development
- url: http://aether.local:8000
  description: Home network